+++ CHANGELOG.md
@@ -23,7 +23,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/google/meridian/compare/v0.6.0...HEAD
+* Make `get_r_hat` public.
 
 ## [0.6.0] - 2024-08-20
 
@@ -60,3 +60,4 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 [0.4.0]: https://github.com/google/meridian/releases/tag/v0.4.0
 [0.5.0]: https://github.com/google/meridian/releases/tag/v0.5.0
 [0.6.0]: https://github.com/google/meridian/releases/tag/v0.6.0
+[Unreleased]: https://github.com/google/meridian/compare/v0.6.0...HEAD